:hammer: Move sourceExclusions to desktop-only config and use DesktopConfigManager
- Remove sourceExclusions from shared AppConfig interface (mobile can't implement source control)
- Delete SourceExclusionService interface from commonMain
- Consolidate all logic into DesktopSourceExclusionService with DesktopConfigManager
- Update SourceControlContentView to use DesktopConfigManager directly (no casting)
- Update all pasteboard services to use DesktopSourceExclusionService type
