+++ app/src/desktopMain/kotlin/com/crosspaste/paste/DesktopSourceExclusionService.kt
@@ -1,7 +1,54 @@
 package com.crosspaste.paste
 
-import com.crosspaste.config.CommonConfigManager
+import com.crosspaste.config.DesktopConfigManager
+import com.crosspaste.utils.getJsonUtils
 
 class DesktopSourceExclusionService(
-    override val configManager: CommonConfigManager,
-) : SourceExclusionService
+    private val configManager: DesktopConfigManager,
+) {
+
+    private fun getSourceExclusions(): String = configManager.getCurrentConfig().sourceExclusions
+
+    fun isExcluded(source: String?): Boolean {
+        if (source == null) return false
+        val exclusions: List<String> =
+            getJsonUtils().JSON.decodeFromString<List<String>>(
+                getSourceExclusions(),
+            )
+        return source in exclusions
+    }
+
+    fun addExclusion(source: String) {
+        val jsonUtils = getJsonUtils()
+        val current: MutableList<String> =
+            jsonUtils.JSON
+                .decodeFromString<List<String>>(
+                    getSourceExclusions(),
+                ).toMutableList()
+        if (source !in current) {
+            current.add(source)
+            configManager.updateConfig(
+                "sourceExclusions",
+                jsonUtils.JSON.encodeToString(current),
+            )
+        }
+    }
+
+    fun removeExclusion(source: String) {
+        val jsonUtils = getJsonUtils()
+        val current: List<String> =
+            jsonUtils.JSON
+                .decodeFromString<List<String>>(
+                    getSourceExclusions(),
+                ).filter { it != source }
+        configManager.updateConfig(
+            "sourceExclusions",
+            jsonUtils.JSON.encodeToString(current),
+        )
+    }
+
+    fun getExclusions(): List<String> =
+        getJsonUtils().JSON.decodeFromString<List<String>>(
+            getSourceExclusions(),
+        )
+}
