+++ shared/src/commonMain/kotlin/com/crosspaste/config/AppConfig.kt
@@ -21,7 +21,6 @@ interface AppConfig {
     val enableSkipPreLaunchPasteboardContent: Boolean
     val lastPasteboardChangeCount: Int
     val enablePasteboardListening: Boolean
-    val sourceExclusions: String
 
     // MB
     val maxBackupFileSize: Long
